+++ IsekaiMod/Content/Classes/IsekaiProtagonist/IsekaiProtagonistClass.cs
@@ -30,6 +30,8 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist
 
         public static void Add()
         {
+            // TODO: rework training arc feats
+
             // TODO: add scaling natural armor, strength and dexterity (like animal companion) to deathsnatcher
             // TODO: add vampiric drain spell for isekai vampire heritage
             // TODO: Load localisation instead of hardcoded strings